+++ CHANGELOG.md
@@ -23,3 +23,9 @@
 * Replaced mysql_secure_installation with bash/expect routines into the mysql CLI client, sending the relevant queries.  This eliminates automation-breaking user prompts.
 
 * Added MediaWiki v1.30.0 installation w/prerequisites.  Verified functional.  After the script runs users should be able to browse to http://Server.IP.Address.Here/mediawiki 
+
+# 2018-03-15
+
+* Added muDBSchema, muDBUser, and muDBPass variables and corresponding CLI parameters for -mus=SchemaName, -muu=UserName, -mup=UserPassword which creates a MySQL Database and sets up a privileged User with the specified password.  The script will inform the user of exactly what these values are (Schema, Username, Password) when complete.
+
+* Added randPass() function to script, which allows better/more secure default password values.  Default passwords for MySQL Root, MU DB User, and MediaWiki Admin are now randomly generated for each script run.  Values may still be overridden by CLI Parameters and the old (simplified) passwords are still there - you can always comment out the single line values that call on the randPass generator to keep the simple default passwords.